pyproject.toml
[tool.poetry]
name = "django-ninja-boilerplate"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "~3.9"
django = "^4.2"
uvicorn = {extras = ["standard"], version = "^0.23.0"}
gunicorn = "^20.1.0"
psycopg2 = "^2.9.5"
django-ninja = "^1.1.0"
whitenoise = "^6.6.0"

[tool.poetry.group.dev.dependencies]
mypy = "^0.910"
flake8 = "^4.0.1"
isort = "^5.9.3"
django-extensions = "^3.2.3"
pygraphviz = "^1.11"
coverage = "^7.3.3"
django-silk = "^5.0.4"
factory-boy = "^3.3.0"
black = "^23.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.isort]
line_length = 119

[tool.mypy]
ignore_missing_imports = true